# openapi.yaml
openapi: 3.0.3
info:
  title: Nexus Teammate API
  version: 0.1.0
  description: |
    MVP OpenAPI draft for Nexus Teammate.
    Tech stack: React + Midwayjs + MySQL.
servers:
  - url: /api/v1
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: Get project list
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Keyword'
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/Project'
    post:
      tags: [Projects]
      summary: Create project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '200':
          description: Project created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Project'
  /projects/{projectId}/boards:
    get:
      tags: [Boards]
      summary: Get boards in project
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Board list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Board'
    post:
      tags: [Boards]
      summary: Create board
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardRequest'
      responses:
        '200':
          description: Board created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Board'
...
  /automation/rules:
    get:
      tags: [Automation]
      summary: Get automation rule list
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: projectId
          in: query
          schema:
            type: integer
            format: int64
        - name: boardId
          in: query
          schema:
            type: integer
            format: int64
        - name: eventType
          in: query
          schema:
            type: string
        - name: isEnabled
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Rule list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/AutomationRule'
    post:
      tags: [Automation]
      summary: Create automation rule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAutomationRuleRequest'
      responses:
        '200':
          description: Rule created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AutomationRule'
  /automation/rules/validate:
    post:
      tags: [Automation]
      summary: Validate automation DSL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateAutomationRuleRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RuleValidationResult'
  /agents:
    get:
      tags: [Agents]
      summary: Get agent list
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Keyword'
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/AgentDefinition'
    post:
      tags: [Agents]
      summary: Create agent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '200':
          description: Agent created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AgentDefinition'
  /agents/{agentId}/execute:
    post:
      tags: [Agents]
      summary: Execute agent manually
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteAgentRequest'
      responses:
        '200':
          description: Execution queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AgentExecutionQueuedResponse'
  /agents/executions/{executionId}:
    get:
      tags: [Agents]
      summary: Get agent execution detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ExecutionId'
      responses:
        '200':
          description: Execution detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AgentExecutionDetail'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    AgentId:
      name: agentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ExecutionId:
      name: executionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    ApiResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          nullable: true
        requestId:
          type: string
    PaginatedResponse:
      type: object
      required: [code, message, data, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
          properties:
            list:
              type: array
              items: {}
            pagination:
              $ref: '#/components/schemas/Pagination'
        requestId:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...
    Project:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
        ownerId:
          type: integer
          format: int64
        status:
          type: string
    CreateProjectRequest:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        ownerId:
          type: integer
          format: int64
    Board:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        name:
          type: string
        viewType:
          type: string
        status:
          type: string
    CreateBoardRequest:
      type: object
      required: [name, viewType]
      properties:
        name:
          type: string
        viewType:
          type: string
          enum: [kanban, list]
        templateId:
          type: integer
          format: int64
          nullable: true
...
    AutomationAction:
      type: object
      properties:
        type:
          type: string
        agentKey:
          type: string
        to:
          type: string
    AutomationRule:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
          nullable: true
        boardId:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        ruleKey:
          type: string
        eventType:
          type: string
        priority:
          type: integer
        isEnabled:
          type: boolean
        conditionDsl:
          type: object
          additionalProperties: true
        actionDsl:
          type: array
          items:
            $ref: '#/components/schemas/AutomationAction'
    CreateAutomationRuleRequest:
      type: object
      required: [name, ruleKey, eventType, priority, conditionDsl, actionDsl]
      properties:
        projectId:
          type: integer
          format: int64
          nullable: true
        boardId:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        ruleKey:
          type: string
        eventType:
          type: string
        priority:
          type: integer
        isEnabled:
          type: boolean
        conditionDsl:
          type: object
          additionalProperties: true
        actionDsl:
          type: array
          items:
            $ref: '#/components/schemas/AutomationAction'
    ValidateAutomationRuleRequest:
      type: object
      required: [eventType, conditionDsl, actionDsl]
      properties:
        eventType:
          type: string
        conditionDsl:
          type: object
          additionalProperties: true
        actionDsl:
          type: array
          items:
            $ref: '#/components/schemas/AutomationAction'
    RuleValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
    AgentDefinition:
      type: object
      properties:
        id:
          type: integer
          format: int64
        agentKey:
          type: string
        name:
          type: string
        category:
          type: string
        status:
          type: string
        modelConfig:
          type: object
          additionalProperties: true
        timeoutSeconds:
          type: integer
        version:
          type: integer
    CreateAgentRequest:
      type: object
      required:
        [agentKey, name, category, status, promptTemplate, inputSchema, modelConfig, timeoutSeconds]
      properties:
        agentKey:
          type: string
        name:
          type: string
        category:
          type: string
        status:
          type: string
        promptTemplate:
          type: string
        inputSchema:
          type: object
          additionalProperties: true
        outputSchema:
          type: object
          additionalProperties: true
        toolPolicy:
          type: object
          additionalProperties: true
        modelConfig:
          type: object
          additionalProperties: true
        timeoutSeconds:
          type: integer
    ExecuteAgentRequest:
      type: object
      required: [input, triggerType]
      properties:
        taskId:
          type: integer
          format: int64
          nullable: true
        input:
          type: object
          additionalProperties: true
        triggerType:
          type: string
    AgentExecutionQueuedResponse:
      type: object
      properties:
        executionId:
          type: integer
          format: int64
        status:
          type: string
    AgentExecutionLog:
      type: object
      properties:
        logType:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    AgentExecutionDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
          nullable: true
        agentId:
          type: integer
          format: int64
        status:
          type: string
        triggerType:
          type: string
        inputPayload:
          type: object
          additionalProperties: true
        outputPayload:
          type: object
          additionalProperties: true
          nullable: true
        errorPayload:
          type: object
          additionalProperties: true
          nullable: true
        costAmount:
          type: number
          format: double
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AgentExecutionLog'
...
security:
  - bearerAuth: []